+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+* Added support for overriding the Event ID when updating, for recurring events
+
 ### Changed
 
 ### Deprecated